+++ src/finger.h
@@ -8,10 +8,10 @@ public:
   int id;
   double score;
   double dexterity;
-  std::pair<int, int> home_position;
-  std::pair<int, int> current_position;
+  std::pair<double, double> home_position;
+  std::pair<double, double> current_position;
 
-  Finger(int id, std::pair<int, int> home_pos, double dex);
+  Finger(int id, std::pair<double, double> home_pos, double dex);
 
   void press(const Key& key);
   void returnToHome();
